app/requirements.txt
streamlit
opencv-python
numpy==1.26.1
pillow==10.1.0
prometheus-client
pydantic-settings==2.0.3
tritonclient[all]==2.34.0